# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Api",
        "description": "Descrição da Sua API",
        "contact": {
            "email": "willnmafra@gmail.com"
        },
        "license": {
            "name": "MIT License"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lease": {
            "get": {
                "tags": [
                    "Leases"
                ],
                "summary": "Get list of leases",
                "operationId": "4cf9a349d2a09f5742db6d8d341e114f",
                "parameters": [
                    {
                        "name": "carmodel_columns",
                        "in": "query",
                        "description": "(optional)Columns to select from realationated tables. ex: /api/car?car_model_id==:1",
                        "schema": {
                            "properties": {
                                "costumer_name": {
                                    "type": "text",
                                    "example": "ilike:%user%"
                                },
                                "car_model_id": {
                                    "type": "text",
                                    "example": "=:5"
                                },
                                "car_plate": {
                                    "type": "text",
                                    "example": "=:MGQ3058"
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "description": "Filters for leases",
                        "schema": {
                            "properties": {
                                "start_date": {
                                    "type": "text",
                                    "example": "=:2024-01-01"
                                },
                                "expected_end_date": {
                                    "type": "text",
                                    "example": "=:2024-12-31"
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Number of items per page for pagination",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Leases"
                ],
                "summary": "Create a new lease",
                "operationId": "c7914fba37c9e83b150f8bad2d2d1dd2",
                "requestBody": {
                    "description": "Lease data to be created",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Car ID",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "start_date": {
                                        "description": "Lease Start Date",
                                        "type": "string",
                                        "example": "2024-01-01 00:00:00"
                                    },
                                    "expected_end_date": {
                                        "description": "Lease Expected End Date",
                                        "type": "string",
                                        "example": "2024-01-01 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lease created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/lease/{leaseId}": {
            "get": {
                "tags": [
                    "Leases"
                ],
                "summary": "Get details of a specific lease",
                "operationId": "26550988ff3a07df33f3da00022b12d1",
                "parameters": [
                    {
                        "name": "leaseId",
                        "in": "path",
                        "description": "ID of the lease",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A json with lease data"
                    },
                    "404": {
                        "description": "Lease not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Leases"
                ],
                "summary": "Update a lease",
                "operationId": "b9d9b899102afd1f9df5230d4b3454d5",
                "parameters": [
                    {
                        "name": "leaseId",
                        "in": "path",
                        "description": "ID of the lease",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "method for the request",
                                        "type": "string",
                                        "example": "patch"
                                    },
                                    "final_km": {
                                        "description": "Final KM of the car",
                                        "type": "integer",
                                        "example": "20000"
                                    },
                                    "actual_end_date": {
                                        "description": "Actual end date of the lease",
                                        "type": "text",
                                        "example": "2024-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Lease updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lease not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leases"
                ],
                "summary": "Delete a lease",
                "operationId": "fd24b600ee2897d3dcfdac810b0870cb",
                "parameters": [
                    {
                        "name": "leaseId",
                        "in": "path",
                        "description": "ID of the lease",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Lease deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lease not found"
                    }
                }
            }
        }
    }
}